new(demo): refactor gradients to Codesandbox example
